track the geometric information of PacMan mouth in the PacMan sprite

+++ src/main/java/com/ahmedaraby/game/pacman/config/intConfigs/ConfigsEx.java
@@ -84,43 +84,43 @@ public class ConfigsEx {
         return PACMAN_MOUTH_ANIMATION_COMPLETE_DIST() * PACMAN_MOUTH_ANIMATION_PERCENTAGES().get(1);
     }
 
-    public double PACMAN_MOUTH_OPEN_ARC_EXTENT_DEG() {
+    public int PACMAN_MOUTH_OPEN_ARC_EXTENT_DEG() {
         return pacman.getMouth().getOpen().getArcExtentDeg();
     }
 
-    public double PACMAN_MOUTH_OPEN_RIGHT_START_ANGLE() {
+    public int PACMAN_MOUTH_OPEN_RIGHT_START_ANGLE() {
         return pacman.getMouth().getOpen().getRightStartAngle();
     }
 
-    public double PACMAN_MOUTH_OPEN_UP_START_ANGLE() {
+    public int PACMAN_MOUTH_OPEN_UP_START_ANGLE() {
         return pacman.getMouth().getOpen().getUpStartAngle();
     }
 
-    public double PACMAN_MOUTH_OPEN_LEFT_START_ANGLE() {
+    public int PACMAN_MOUTH_OPEN_LEFT_START_ANGLE() {
         return pacman.getMouth().getOpen().getLeftStartAngle();
     }
 
-    public double PACMAN_MOUTH_OPEN_DOWN_START_ANGLE() {
+    public int PACMAN_MOUTH_OPEN_DOWN_START_ANGLE() {
         return pacman.getMouth().getOpen().getDownStartAngle();
     }
 
-    public double PACMAN_MOUTH_CLOSED_ARC_EXTENT_DEG() {
+    public int PACMAN_MOUTH_CLOSED_ARC_EXTENT_DEG() {
         return pacman.getMouth().getClosed().getArcExtentDeg();
     }
 
-    public double PACMAN_MOUTH_CLOSED_RIGHT_START_ANGLE() {
+    public int PACMAN_MOUTH_CLOSED_RIGHT_START_ANGLE() {
         return pacman.getMouth().getClosed().getRightStartAngle();
     }
 
-    public double PACMAN_MOUTH_CLOSED_UP_START_ANGLE() {
+    public int PACMAN_MOUTH_CLOSED_UP_START_ANGLE() {
         return pacman.getMouth().getClosed().getUpStartAngle();
     }
 
-    public double PACMAN_MOUTH_CLOSED_LEFT_START_ANGLE() {
+    public int PACMAN_MOUTH_CLOSED_LEFT_START_ANGLE() {
         return pacman.getMouth().getClosed().getLeftStartAngle();
     }
 
-    public double PACMAN_MOUTH_CLOSED_DOWN_START_ANGLE() {
+    public int PACMAN_MOUTH_CLOSED_DOWN_START_ANGLE() {
         return pacman.getMouth().getClosed().getDownStartAngle();
     }
